+++ crates/astria-sequencer/CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Remove failed promotable instead of inserted transaction during mempool insertion
+  [#2135](https://github.com/astriaorg/astria/pull/2135).
+
 ## [3.0.0-rc.1]
 
 ### Added